Chanded cycle_num type to uint64_t for consistancy

+++ parser/state.h
@@ -2,13 +2,15 @@
 #define STATE_H
 
 #include <string>
+#include <cstdint>
+
 #include "instruction.h"
 
 // class for describing the program state at a specific cycle
 class State {
     private:
 
-        unsigned cycle_num;
+        uint64_t cycle_num;
         Instruction in_fetch;
         Instruction in_decode;
         Instruction in_execute;
@@ -18,7 +20,7 @@ class State {
     public:
 
         State(Instruction f, Instruction d, Instruction e, Instruction m, Instruction wb);
-        unsigned get_cycle_num();
+        uint64_t get_cycle_num();
         std::string get_in_fetch();
         std::string get_in_decode();
         std::string get_in_execute();
